Clean-up some of the server
