+++ server/index.ts
@@ -8,6 +8,7 @@ import path from 'path';
 import send from 'koa-send';
 
 import { getEnv } from './env';
+import { applyMiddleware } from './middleware';
 
 const app = new Koa();
 const router = new Router();
@@ -20,38 +21,16 @@ socket.on('connect', socket => {
   console.log(`Got a bite!`);
 });
 
-// logger
-
-app.use(async (ctx, next) => {
-  await next();
-  const rt = ctx.response.get('X-Response-Time');
-  console.log(`${ctx.method} ${ctx.url} - ${rt}`);
-});
-
-// x-response-time
-
-app.use(async (ctx, next) => {
-  const start = Date.now();
-  await next();
-  const ms = Date.now() - start;
-  ctx.set('X-Response-Time', `${ms}ms`);
-});
-
-// Static file server:
-app.use(async (ctx, next) => {
-  if (ctx.path === '/') {
-    await send(ctx, '/dist/index.html', {
-      root: path.resolve(__dirname, '..')
-    });
-  }
-  if (ctx.path.startsWith('/dist/')) {
-    await send(ctx, ctx.path, { root: path.resolve(__dirname, '..') });
-  }
-  await next();
-});
-
+applyMiddleware(app);
 app.use(router.routes());
 app.use(router.allowedMethods());
+
+router.post('/player', (ctx, next) => {
+  const { name } = ctx.request.body;
+  console.log(`Registering ${name}`);
+
+  ctx.status = 200;
+});
 
 app.listen(env.port, () => {
   console.log(`Listening on port ${env.port}`);
